requirements.txt
# Core ML
torch>=2.0.0
torch-geometric>=2.4.0
torch-scatter
torch-sparse
torch-cluster

# Protein / Bio
biopython>=1.81
fair-esm>=2.0.0

# Chemistry
rdkit-pypi>=2022.9.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.2.0
h5py>=3.8.0

# Visualization & logging
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Utilities
tqdm>=4.65.0
requests>=2.28.0
joblib>=1.2.0

# Structure analysis
pymol-open-source>=2.5.0; sys_platform == 'linux'

# API access
xmltodict>=0.13.0
